move to own trait
