+++ tests/src/FunctionalJavascript/Integration/InlineEntityFormTest.php
@@ -4,6 +4,7 @@ namespace Drupal\Tests\thunder\FunctionalJavascript\Integration;
 
 use Drupal\Tests\thunder\FunctionalJavascript\ThunderJavascriptTestBase;
 use Drupal\Tests\thunder\FunctionalJavascript\ThunderFormFieldTestTrait;
+use Drupal\Tests\thunder\FunctionalJavascript\ThunderJavascriptTrait;
 
 /**
  * Test for update hook changes.
@@ -15,6 +16,7 @@ use Drupal\Tests\thunder\FunctionalJavascript\ThunderFormFieldTestTrait;
 class InlineEntityFormTest extends ThunderJavascriptTestBase {
 
   use ThunderFormFieldTestTrait;
+  use ThunderJavascriptTrait;
 
   /**
    * Test saving collapsed gallery paragraph.
